backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
google-auth==2.26.1
google-auth-oauthlib==1.2.0
python-jose[cryptography]==3.3.0

# Firebase
firebase-admin==6.4.0

# MQTT
paho-mqtt==1.6.1

# Machine Learning
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
joblib==1.3.2

# SMS Alerts
twilio==8.11.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
